+++ resources/js/components/adminpanel/FooterPanel.vue
